+++ src/main/java/ru/Hmelev/tm/service/TaskService.java
@@ -3,58 +3,51 @@ package ru.Hmelev.tm.service;
 import ru.Hmelev.tm.entity.Task;
 import ru.Hmelev.tm.repository.TasksRepository;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.Iterator;
-import java.util.List;
+import java.util.*;
 
 public class TaskService extends Service {
-    private TasksRepository taskRepository;
+    private final TasksRepository taskRepository;
     private Task task;
 
-    private List<Task> list = new ArrayList<>();
+    private final List<Task> list = new ArrayList<>();
 
     public TaskService(TasksRepository taskRepository) {
         this.taskRepository = taskRepository;
     }
 
     public void createTask(String id, String name, String description, Date startDate, Date finishDate, String idProject) {
-        task = new Task(id, name, description, startDate, finishDate, idProject);
-        taskRepository.persist(id, task);
+        if (id != null && !id.isEmpty() && name != null && !name.isEmpty() && description != null && !description.isEmpty()
+                && startDate != null && finishDate != null && idProject != null && !idProject.isEmpty()) {
+            task = new Task(id, name, description, startDate, finishDate, idProject);
+            taskRepository.persist(id, task);
+        }
     }
 
     public void clearTask() {
         taskRepository.removeAll();
     }
 
-    public void listTask() {
-        for (Task task : taskRepository.findAll()) {
-            showTask(task.getId());
-        }
-
+    public Collection<Task> findAllTasks() {
+        return taskRepository.findAll();
     }
 
     public void editTask(String id, String name, String description, Date startDate, Date finishDate, String idProject) {
-        task = taskRepository.findOne(id);
-        task.setName(name);
-        task.setDescription(description);
-        task.setStartDate(startDate);
-        task.setFinishDate(finishDate);
-        task.setIdProject(idProject);
+        if (id != null && !id.isEmpty() && name != null && !name.isEmpty() && description != null && !description.isEmpty()
+                && startDate != null && finishDate != null && idProject != null && !idProject.isEmpty()) {
+            task = taskRepository.findOne(id);
+            task.setName(name);
+            task.setDescription(description);
+            task.setStartDate(startDate);
+            task.setFinishDate(finishDate);
+            task.setIdProject(idProject);
+        }
     }
 
-    public void showTask(String id) {
-        Task task = taskRepository.findOne(id);
-        if (task != null) {
-            System.out.println(
-                    "[ ID = " + task.getId()
-                            + "; Name = " + task.getName()
-                            + "; Description = " + task.getDescription()
-                            + "; Start date = " + DEFAULT_DATE_FORMAT.format(task.getStartDate())
-                            + "; Finish date = " + DEFAULT_DATE_FORMAT.format(task.getFinishDate())
-                            + "; idProject = " + task.getIdProject()
-                            + " ]");
+    public Task findTask(String id) {
+        if (id != null && !id.isEmpty()) {
+            return taskRepository.findOne(id);
         }
+        return null;
     }
 
     public void removeTask(String id) {
@@ -62,24 +55,29 @@ public class TaskService extends Service {
     }
 
     public List<Task> listTaskIdProject(String idProject) {
-        list.addAll(taskRepository.findAll());
-        Iterator<Task> it = list.iterator();
-        while (it.hasNext()) {
-            task = it.next();
-            if (!task.getIdProject().equals(idProject)) {
-                it.remove();
+        if (idProject != null && !idProject.isEmpty()) {
+            list.addAll(taskRepository.findAll());
+            Iterator<Task> it = list.iterator();
+            while (it.hasNext()) {
+                task = it.next();
+                if (!task.getIdProject().equals(idProject)) {
+                    it.remove();
+                }
             }
+            return list;
         }
-        return list;
+        return null;
     }
 
     public void listTaskNoIdProject(String idProject) {
-        list.addAll(taskRepository.findAll());
-        Iterator<Task> it = list.iterator();
-        while (it.hasNext()) {
-            task = it.next();
-            if (task.getIdProject().equals(idProject)) {
-                taskRepository.remove(task.getId());
+        if (idProject != null && !idProject.isEmpty()) {
+            list.addAll(taskRepository.findAll());
+            Iterator<Task> it = list.iterator();
+            while (it.hasNext()) {
+                task = it.next();
+                if (task.getIdProject().equals(idProject)) {
+                    taskRepository.remove(task.getId());
+                }
             }
         }
     }
